adds better XP system and a global top command which integrates all stats

+++ internal/bot-service/data/models/user.go
@@ -8,6 +8,6 @@ type User struct {
 	CurrentCircle     string
 	CurrentInnerOrder *int
 	CurrentLevel      int
-	CurrentExperience int
+	CurrentExperience float64
 	CreatedAt         *int64
 }
